// 01-starting-project/src/app/model/AnnaulData.ts
export interface AnnaulData {
  year: number;
  interest: number;
  valueEndOfYear: number;
  annualInvestment: number;
  totalInterest: number;
  totalAmountInvested: number;
}

export  interface UserInputData {
  initialInvestment: number ,
    annualInvestment: number ,
  expectedReturn: number ,
  duration: number
}